created hook for duplicate code
