+++ solidstart-tanstackquery-tailwind-modules/src/routes/[owner]/[name]/tree/[branch]/[...path]/index.tsx
@@ -1,49 +1,13 @@
-import { createQuery } from '@tanstack/solid-query';
-import { createEffect, createSignal } from 'solid-js';
 import { useParams } from 'solid-start';
 import { BranchNavigation } from '~/components/BranchNavigation';
 import FileExplorer from '~/components/FileExplorer';
 import { RepoHeader } from '~/components/RepoHeader';
-import getRepoInfo from '~/services/get-repo-info';
-import { Info } from '~/types/repo-info-type';
 import styles from '../../../../style.module.css';
+import useGetRepoInfo from '~/hooks/useGetRepoInfo';
 
 const RepoTree = () => {
   const params = useParams();
-  const [branch, setBranch] = createSignal(params.branch);
-  const [info, setInfo] = createSignal<Info>({
-    isPrivate: false,
-    visibility: '',
-    forkCount: 0,
-    description: '',
-    homepageUrl: '',
-    stargazerCount: 0,
-    watcherCount: 0,
-    openIssueCount: 0,
-    topics: [],
-    isOrg: false,
-    openPullRequestCount: 0,
-  });
-
-  const isOwnerAndNameValid =
-    typeof params.owner === 'string' && typeof params.name === 'string';
-
-  const repoInfo = createQuery(
-    () => [`repository-info_${params.owner}_${params.name}`],
-    () =>
-      getRepoInfo(
-        isOwnerAndNameValid
-          ? { owner: params.owner, name: params.name }
-          : { owner: '', name: '' }
-      )
-  );
-
-  createEffect(() => {
-    if (repoInfo.isSuccess && !repoInfo.isLoading && repoInfo.data) {
-      setInfo(repoInfo.data.info);
-      setBranch(repoInfo.data.branch || params.branch);
-    }
-  });
+  const [info, branch,] = useGetRepoInfo();
 
   return (
     <div class={styles.wrapper}>
